Rename 'categories' to 'topics'
Even though the underlying option names are still 'podcast_category_*',
we should still rename 'category' to 'topic' in other parts of the code:

- It fixes a lint error.
- It's consistent with the UI this way.
- We already use "category" to mean something else in this same settings
  screen, and overloading the meaning is confusing.
